+++ Ligo/Modules/Professions/Events/GameLoop/UpdateTicked/SlimeDeflationUpdateTickedEvent.cs
@@ -23,7 +23,10 @@ internal sealed class SlimeDeflationUpdateTickedEvent : UpdateTickedEvent
     /// <inheritdoc />
     protected override void OnUpdateTickedImpl(object? sender, UpdateTickedEventArgs e)
     {
-        var undeflated = GreenSlime_Piped.Values.Select(pair => pair.Key).ToArray();
+        var undeflated = GreenSlime_Piped.Values
+            .Select(pair => pair.Value)
+            .Where(piped => piped.PipeTimer <= 0 && piped.Instance.Scale > piped.OriginalScale)
+            .ToArray();
         if (undeflated.Length == 0)
         {
             this.Disable();