+++ Ligo/Modules/Arsenal/CHANGELOG.md
@@ -1,5 +1,5 @@
 ﻿# Ligo Arsenal (Slingshots) Change Log
 
-## 1.0.0 (Unreleased)
+## 0.9.0 (Unreleased)
 
 * Initial Version