+++ Ligo/Modules/Tools/CHANGELOG.md
@@ -1,6 +1,6 @@
 ﻿# Ligo Tools Change Log
 
-## 1.0.0 (Unreleased)
+## 0.9.0 (Unreleased)
 
 ### Added