+++ Ligo/Modules/Tweex/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Ligo Tweaks Change Log
 
-## 1.0.0 (Unreleased)
+## 0.9.0 (Unreleased)
 
 ### Added